requirements.txt
configparser==3.5.0
Django==1.11.5
django-markdown-deux==1.0.5
django-suit==2.0a1
markdown2==2.3.4
mysqlclient==1.3.12
pkg-resources==0.0.0
pytz==2017.2
